requirements.txt
# Job scraping 
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
python-dotenv==1.0.0

# API Integration
openai==1.12.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0

# Data processing
pandas>=2.0.0
python-docx==1.1.0
docxtpl==0.16.7

# Web interface (optional)
flask==2.3.3 